🗑🧹Garbage Collection and File Structure Cleanse
Lots of moving around of components and deleting of unused components and css
