+++ src/components/Globals/section-header.css
@@ -17,4 +17,5 @@ header p {
   color: #743100;
   font-size: 1.5rem;
   font-weight: bold;
+  font-family: "Nunito", sans-serif;
 }
