+++ src/components/HomePage/HeroSection.js
@@ -1,6 +1,5 @@
 import { useState, useEffect } from "react";
-import SearchBar from "../SearchBar";
-import HeroPicture from "./images/yosemitevalley (1).png";
+import SearchBar from "../UtilityComponents/SearchBar";
 import "./section-hero.css";
 import HeroPicture from "./images/yosemitevalley.png";
