--- src/components/HomePage/header.css
@@ -1,133 +1,0 @@
-/*Desktop View*/
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0px 40px;
-  height: 80px;
-  max-width: 64rem;
-  margin: 0 auto;
-}
-.logo-nav {
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-}
-.logo-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.logo {
-  width: 200px;
-  height: auto;
-}
-.nav-options {
-  padding-left: 25px;
-  display: grid;
-  grid-template-columns: repeat(3, auto);
-  grid-gap: 50px;
-  list-style-type: none;
-}
-.mobile-option {
-  display: none;
-}
-
-.option :hover {
-  /* color: white; */
-}
-.signin-up {
-  display: flex;
-  padding: 0px 5px;
-  list-style-type: none;
-}
-.sign-in {
-  padding-right: 50px;
-  align-self: center;
-}
-.sign-in :hover {
-  /* color: white; */
-}
-.signup-btn {
-  padding: 10px 10px;
-  height: 2.4rem;
-  border-radius: 30px;
-  opacity: 0.8;
-  transition: 0.3s;
-  color: #fafafa;
-  background-color: #e2aa60;
-  cursor: pointer;
-  align-items: center;
-}
-.signup-btn:hover {
-  opacity: 1;
-}
-.mobile-menu {
-  display: none;
-}
-@media (max-width: 684px) {
-  /*Mobile View */
-  .header {
-    padding: 0px 10px;
-  }
-  .nav-options {
-    display: flex;
-    width: 100%;
-    height: 350px;
-    position: absolute;
-    top: 80px;
-    left: -100%;
-    opacity: 0;
-    transition: all 0.5s ease;
-    flex-direction: column;
-    list-style-type: none;
-    grid-gap: 0px;
-  }
-  .nav-options.active {
-    left: 0;
-    opacity: 1;
-    transition: all 0.5s ease;
-    z-index: 1;
-    align-content: center;
-    padding-left: 0px;
-  }
-
-  .menu-icon {
-    width: 45px;
-    height: 45px;
-  }
-  .option {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 10vw;
-    padding: 30px 0px;
-    background: #fefef8;
-  }
-  .sign-up {
-    border-radius: 3px;
-    color: #743100;
-    padding: 20px 0px;
-    width: 80%;
-    align-self: center;
-    border-radius: 3px;
-    background-color: #743100;
-    color: #fefef8;
-    border-radius: 89px;
-    padding: 10px 0px;
-    width: 30%;
-    text-align: center;
-  }
-
-  .signin-up {
-    display: none;
-  }
-
-  .mobile-menu {
-    display: block;
-  }
-}
-
-.header-link {
-  color: #9b5933;
-}
