--- src/components/HomePage/section-footer.css
@@ -1,81 +1,0 @@
-footer {
-  grid-column: 1 / 13;
-  grid-row: 22 / 26;
-  background: linear-gradient(#743100, #8e3a24) !important;
-  color: #fafafa;
-  padding: 60px 0;
-  margin: 1rem;
-}
-
-.container {
-  width: 100%;
-  max-width: 1100px;
-  padding: 0 20px;
-  margin: 0 auto;
-}
-
-ul {
-  padding: 0;
-}
-
-li {
-  list-style: none;
-  line-height: 2;
-  text-align: left !important;
-}
-
-a {
-  color: #fafafa;
-  text-decoration: none;
-}
-
-a:hover {
-  text-decoration: underline;
-}
-
-.footer-content {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  flex-wrap: wrap;
-}
-
-.items {
-  flex-basis: 20%;
-}
-
-.items .logo {
-  text-transform: capitalize;
-  font-size: 1.4em;
-  color: #fafafa;
-}
-
-.items p {
-  margin: 1.2em 0;
-}
-
-.items h3 {
-  position: relative;
-  text-align: left !important;
-  font-size: 1.3em;
-  margin-bottom: 1rem;
-  color: #fafafa;
-}
-.items h3::after {
-  content: "";
-  position: absolute;
-  width: 40px;
-  height: 3px;
-  background: #fafafa;
-  bottom: -4px;
-  left: 0;
-}
-
-hr {
-  opacity: 0.3;
-}
-
-footer .end {
-  text-align: center;
-  padding: 20px 0 0 0;
-}
